# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "Erro": {
        "description": "O conteúdo do campo `erro`.",
        "properties": {
          "code": {
            "description": "Código estável; é nele que o código do cliente decide",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "O que é específico deste erro, como o campo que falhou",
            "title": "Details"
          },
          "message": {
            "description": "Mensagem em português, para mostrar ao aluno",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Id da requisição, o mesmo que aparece no log",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "Erro",
        "type": "object"
      },
...
      "ModuloDetalhe": {
        "properties": {
          "atividades": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AtividadeResumo"
            },
            "title": "Atividades",
            "type": "array"
          },
          "bloqueado": {
            "default": true,
            "title": "Bloqueado",
            "type": "boolean"
          },
          "concluido": {
            "default": false,
            "title": "Concluido",
            "type": "boolean"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ordem": {
            "title": "Ordem",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/StatusProgresso",
            "default": "bloqueado"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          }
        },
        "required": [
          "id",
          "ordem",
          "titulo"
        ],
        "title": "ModuloDetalhe",
        "type": "object"
      },
...
      "Pagina_Trilha_": {
        "properties": {
          "itens": {
            "items": {
              "$ref": "#/components/schemas/Trilha"
            },
            "title": "Itens",
            "type": "array"
          },
          "pagina": {
            "description": "Página atual, começando em 1",
            "minimum": 1.0,
            "title": "Pagina",
            "type": "integer"
          },
          "tamanho_pagina": {
            "description": "Quantos itens cabem na página",
            "minimum": 1.0,
            "title": "Tamanho Pagina",
            "type": "integer"
          },
          "total": {
            "description": "Total de itens em todas as páginas",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "total_paginas": {
            "description": "Quantas páginas existem no total",
            "minimum": 0.0,
            "title": "Total Paginas",
            "type": "integer"
          }
        },
        "required": [
          "itens",
          "pagina",
          "tamanho_pagina",
          "total",
          "total_paginas"
        ],
        "title": "Pagina[Trilha]",
        "type": "object"
      },
      "PassoConcluido": {
        "description": "O que volta ao fechar um passo.",
        "properties": {
          "proximo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proximo"
          },
          "xp_ganho": {
            "title": "Xp Ganho",
            "type": "integer"
          },
          "xp_total": {
            "title": "Xp Total",
            "type": "integer"
          }
        },
        "required": [
          "xp_ganho",
          "xp_total"
        ],
        "title": "PassoConcluido",
        "type": "object"
      },
      "PassoDetalhe": {
        "description": "Conteúdo de um passo, com o bastante para a tela 26 se situar.",
        "properties": {
          "atividades": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AtividadeResumo"
            },
            "title": "Atividades",
            "type": "array"
          },
          "bloqueado": {
            "default": true,
            "title": "Bloqueado",
            "type": "boolean"
          },
          "concluido": {
            "default": false,
            "title": "Concluido",
            "type": "boolean"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ordem": {
            "title": "Ordem",
            "type": "integer"
          },
          "proximo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proximo"
          },
          "status": {
            "$ref": "#/components/schemas/StatusProgresso",
            "default": "bloqueado"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          },
          "total_passos": {
            "title": "Total Passos",
            "type": "integer"
          },
          "trilha_nome": {
            "title": "Trilha Nome",
            "type": "string"
          },
          "trilha_slug": {
            "title": "Trilha Slug",
            "type": "string"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          }
        },
        "required": [
          "id",
          "ordem",
          "titulo",
          "trilha_slug",
          "trilha_nome",
          "total_passos"
        ],
        "title": "PassoDetalhe",
        "type": "object"
      },
...
      "RespostaDeErro": {
        "description": "Corpo de qualquer resposta de erro. Existe para o OpenAPI documentá-lo.",
        "properties": {
          "erro": {
            "$ref": "#/components/schemas/Erro"
          }
        },
        "required": [
          "erro"
        ],
        "title": "RespostaDeErro",
        "type": "object"
      },
...
      "StatusProgresso": {
        "enum": [
          "bloqueado",
          "em-andamento",
          "concluido"
        ],
        "title": "StatusProgresso",
        "type": "string"
      },
...
      "Trilha": {
        "properties": {
          "categoria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria"
          },
          "disciplina": {
            "title": "Disciplina",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "periodo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodo"
          },
          "progresso": {
            "default": 0,
            "title": "Progresso",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "total_modulos": {
            "default": 0,
            "title": "Total Modulos",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "nome",
          "disciplina"
        ],
        "title": "Trilha",
        "type": "object"
      },
      "TrilhaDetalhe": {
        "properties": {
          "categoria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "disciplina": {
            "title": "Disciplina",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "modulos": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ModuloDetalhe"
            },
            "title": "Modulos",
            "type": "array"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "periodo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodo"
          },
          "progresso": {
            "default": 0,
            "title": "Progresso",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "total_modulos": {
            "default": 0,
            "title": "Total Modulos",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "nome",
          "disciplina"
        ],
        "title": "TrilhaDetalhe",
        "type": "object"
      },
      "TrilhaRecomendada": {
        "description": "Card da seção de sugestões da tela 24.",
        "properties": {
          "categoria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria"
          },
          "disciplina": {
            "title": "Disciplina",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "motivo": {
            "title": "Motivo",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "periodo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodo"
          },
          "progresso": {
            "default": 0,
            "title": "Progresso",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "total_modulos": {
            "default": 0,
            "title": "Total Modulos",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "nome",
          "disciplina",
          "motivo"
        ],
        "title": "TrilhaRecomendada",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token do /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API do Algorise. As rotas de negócio vivem sob `/api/v1`; `/health`, `/ready` e `/version` ficam fora do prefixo por serem de operação.",
    "title": "Algorise API",
    "version": "nao-versionado"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trilhas": {
      "get": {
        "description": "Trilhas publicadas, paginadas. Com token, cada card vem com o progresso do aluno. Aceita `ordenar_por` em `periodo`, `nome` ou `disciplina`.",
        "operationId": "listarTrilhas",
        "parameters": [
          {
            "description": "Filtra por nome ou disciplina",
            "in": "query",
            "name": "busca",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por nome ou disciplina",
              "title": "Busca"
            }
          },
          {
            "description": "Período letivo; pode repetir para filtrar vários",
            "in": "query",
            "name": "periodo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maximum": 8,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Período letivo; pode repetir para filtrar vários",
              "title": "Periodo"
            }
          },
          {
            "description": "Categoria do card (tela 24)",
            "in": "query",
            "name": "categoria",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Categoria do card (tela 24)",
              "title": "Categoria"
            }
          },
          {
            "description": "Página, começando em 1",
            "in": "query",
            "name": "pagina",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Página, começando em 1",
              "minimum": 1,
              "title": "Pagina",
              "type": "integer"
            }
          },
          {
            "description": "Itens por página (máximo 100)",
            "in": "query",
            "name": "tamanho_pagina",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Itens por página (máximo 100)",
              "maximum": 100,
              "minimum": 1,
              "title": "Tamanho Pagina",
              "type": "integer"
            }
          },
          {
            "description": "Campo de ordenação; cada rota documenta os seus",
            "in": "query",
            "name": "ordenar_por",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Campo de ordenação; cada rota documenta os seus",
              "title": "Ordenar Por"
            }
          },
          {
            "description": "Inverte a ordem",
            "in": "query",
            "name": "decrescente",
            "required": false,
            "schema": {
              "default": false,
              "description": "Inverte a ordem",
              "title": "Decrescente",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pagina_Trilha_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Dados inválidos"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Erro inesperado"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Catálogo de trilhas",
        "tags": [
          "trilhas"
        ]
      }
    },
    "/api/v1/trilhas/recomendadas": {
      "get": {
        "description": "Sugestões conforme o período e o nível informados no onboarding.\n\nFica antes de `/{slug}` de propósito: declarada depois, a rota de detalhe\nengoliria \"recomendadas\" como se fosse um slug.",
        "operationId": "listarRecomendadas",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrilhaRecomendada"
                  },
                  "title": "Response Listarrecomendadas",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Dados inválidos"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Erro inesperado"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trilhas sugeridas para o aluno",
        "tags": [
          "trilhas"
        ]
      }
    },
    "/api/v1/trilhas/{slug}": {
      "get": {
        "operationId": "obterTrilha",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrilhaDetalhe"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Dados inválidos"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Erro inesperado"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trilha com os passos e seus estados",
        "tags": [
          "trilhas"
        ]
      }
    },
    "/api/v1/trilhas/{slug}/iniciar": {
      "post": {
        "description": "Idempotente: quem já começou continua de onde parou.",
        "operationId": "iniciarTrilha",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrilhaDetalhe"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Dados inválidos"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Erro inesperado"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Começa a trilha para o aluno",
        "tags": [
          "trilhas"
        ]
      }
    },
    "/api/v1/trilhas/{slug}/passos/{ordem}": {
      "get": {
        "description": "Recusa com 403 o passo que o aluno ainda não desbloqueou.",
        "operationId": "obterPasso",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ordem",
            "required": true,
            "schema": {
              "title": "Ordem",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassoDetalhe"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Dados inválidos"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Erro inesperado"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Conteúdo de um passo",
        "tags": [
          "trilhas"
        ]
      }
    },
    "/api/v1/trilhas/{slug}/passos/{ordem}/concluir": {
      "post": {
        "description": "Exige ter acertado as atividades corrigíveis do passo. Fechar de novo não paga XP outra vez.",
        "operationId": "concluirPasso",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ordem",
            "required": true,
            "schema": {
              "title": "Ordem",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassoConcluido"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Dados inválidos"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaDeErro"
                }
              }
            },
            "description": "Erro inesperado"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fecha o passo e concede o XP",
        "tags": [
          "trilhas"
        ]
      }
    },
...
  },
...
}